+++ packages/components/src/ProTable/useTable.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { Params, useTableData, stateParams } from './types';
 
 const useTable = (key: string, params: Params = {}): useTableData => {
-  const { formatData, query, SWRConfiguration } = params;
+  const { formatData, query, SWRConfiguration, requestOptions } = params;
 
   // 表单组件实例
   const [form, setForm] = useState<any>(null);
@@ -62,6 +62,7 @@ const useTable = (key: string, params: Params = {}): useTableData => {
   const onSearch = async () => {
     await form.current.onSubmit();
     const isNoError = form.current.getError();
+
     if (Object.keys(isNoError).length === 0) {
       await state.setPageIndex(1);
       state.mutate(false);
@@ -81,6 +82,7 @@ const useTable = (key: string, params: Params = {}): useTableData => {
     updateStore,
     SWRConfiguration,
     form,
+    requestOptions,
     updateForm,
     ...state,
   };